tests: implement ability to run box/engine tests with different allocators
The ability to select an allocator for memtx has been added to
tarantool. To test a new type of allocator, all tests must also
be run with it. Implemented new option, which allows to set allocator
for memtx in tests. If you wan't to choose allocator type for tests,
run test-run.py with --allocator="small" or --allocator="system".
Added job with "system" allocator type for box/engine tests.