memtx: convert some *.c files to *.cc
In the patch with the choice of allocator for memtx, it was
decided to use templates, so we need to change all associated
files from *.c to *.cc. At the same time, changes were made to
ensure c++ compilation: added explicit type casting, refactoring
code with goto, which crosses variable initialization.
Part of #5419