sql: ignore \0 in string passed to Lua-function
Prior to this patch string passed to user-defined Lua-function from SQL
was cropped in case it contains '\0'. At the same time, it wasn't
cropped if it is passed to the function from BOX. After this patch the
string won't be cropped when passed from SQL if it contain '\0'.

Closes #5938

(cherry picked from commit 22e2e4eaad7796f7b0142cc7e603f9fe895f6642)