Updated some of built-ins, added tests
There are number of failing tests in the new 5913 bunch. Number of SQL
built-ins need a heavy rework - such as PRINTF that can accept anything
besides the format string as its first argument, or LUA that according
to the documentation should accept single string only but in fact
iterates over the whole arg list with no limitation.