swim: check types in __serialize methods
In swim Lua code none of the __serialize methods checked the
argument type assuming that nobody would call them directly and
mess with the types. But it happened, and is not hard to fix, so
the patch does it.

The serialization functions are sanitized for the swim object,
swim member, and member event.

Closes #5952

(cherry picked from commit 1d121c126218408f2ab19b22cb1edd2ed1ec8f39)