txm: simlification: don't link stories with clean tuples
Before this patch there was a way to build history of a key when
the terminal element in list is a clean tuple, not its
story. That was a kind of optimization desigend to avoid excess
story creation.

Unfortunately that kind of otimization is possible to use only
for one-index spaces, and even in that case it doesn't work.

Let's remove optimization that doesn't work.

Part of #5628