bps: add successor argument to insertion function
Common insertion function in BSP tree can make a replacement
or insertion.

When a replacement happens, a replaced element is returned via
special agrument.

Let's add another argument of the function for the case of
insetion: an element before which then insertion happened will
be returned via that argument.

Part of #5628