box: introduce graceful_shutdown setting
We need to differentiate sessions that support graceful shutdown and
not.

Added field `graceful_shutdown` to space _session_settings. As default
it is false and session, that support graceful shutdown must set it to
true.

@TarantoolBot document
Title introduce new boolean value in `_session_settings` space -
`graceful_shutdown`. This value must be `true` only if session support
graceful shutdown, so by default it is `false`.
Part of #5924