iproto: add check for user input of the number of iproto threads
At the moment, user can set any number of iproto threads, which leads
to incorrect behavior if the specified number of threads is less than
or equal to zero or too large. Added check for user input of the number
of iproto threads - value must be > 0 and less than or equal to 1000.

Closes #6005

@TarantoolBot document
Title: Add check for user input of the number of iproto threads
Added check for user input of the number of iproto threads - value
must be > 0 and less than or equal to 1000.

(cherry picked from commit d97d1130bc899373b5780a7e8f5d31c449973636)