test: update test-run (fix release_disabled)
The problem that is fixed by this test-run update is the following. A
test that is listed in the 'release_disabled' field in a suite.ini file
of a 'core = app' or a 'core = unittest' test suite was disabled at all,
not only for the release tarantool build. A 'core = tarantool' test
suite was not affected by the problem.

Disabled sql-tap/debug_mode_only.test.lua test, because it does not work
anymore. We plan to revisit this test in the scope of #3694.

See https://github.com/tarantool/test-run/issues/199