memtx: introduce foreach_allocator template walker
It is supposed to generalize workflow with allocators so that apply the
same function (like create()/destroy() etc) for each existing allocator.
Newborn functions are not used yet since we are going to add more
wrappers.

Warning: dark template magic is involved, do not try what you're about
to see at home.

Follow-up #5419