replication: ballot.is_ro -> is_ro_cfg
Rename the member to show its actual meaning. It is not the
real RO state of the instance. Only how it is configured.

It can happen that the instance is read_only = false, but still is
in RO state due to other reasons.

The patch is done in scope of #5613 where the ballot is going to
be extended and used a bit differently in the join-master search
algorithm.

Part of #5613

(cherry picked from commit e4488f47e0b55df98f1e6d6355001ecf060da8ab)