txm: fix transaction preparation
There were several problems that was connected with broken
pointers in tuple history.

A story of a tuple holds two pointers to transaction statements,
one for TX that added this tuple and one for TX that deleted it.

When a statement is prerared (after user's box.commit() call),
the list of stories can be reordered in order to maintain the
required order of stories - inprogress -> prepared -> committed.

During that reorder it is also necessary to maintain correct
pointers to transaction statements.

This patch refactors the preparation and fixes the problem.

Closes #6132
Closes #6021